update：实现文件上传后计入已用空间，修改记录存储空间的数据类型由numeric修改为int8以及更正涉及该变动的代码

+++ src/main/java/com/zoi/drive/entity/dto/UserCheckin.java
@@ -43,7 +43,7 @@ public class UserCheckin implements Serializable {
     private Date lastCheckin;
 
     @Schema(description = "签到奖励")
-    private BigDecimal checkinReward;
+    private long checkinReward;
 
     @Schema(description = "当天为止连续签到天数")
     private Integer checkinConsecutive;
